src/scrapers/web-scraper.py
import scrapy
from typing import Dict, List, Any
from .base_scraper import BaseScraper
from bs4 import BeautifulSoup

class WebScraper(BaseScraper):
    def __init__(self, config: Dict[str, Any]):
        self.config = config
        
    def scrape(self, source: str) -> List[Dict[str, Any]]:
        """Scrape web content using Scrapy."""
        # Implementation of web scraping logic
        # This is a placeholder and should be replaced with actual scraping code
        return [{"url": source, "html": "<html><body>Sample content</body></html>"}]
    
    def clean(self, data: Dict[str, Any]) -> Dict[str, Any]:
        """Clean HTML content."""
        soup = BeautifulSoup(data['html'], 'html.parser')
        cleaned_data = {
            "url": data['url'],
            "text": soup.get_text(),
            "title": soup.title.string if soup.title else ""
        }
        return cleaned_data
