requirements.txt

scrapy==2.5.1
neo4j==4.4.3
transformers==4.21.0
langchain==0.0.184
apache-airflow==2.3.3
pyyaml==6.0
beautifulsoup4==4.11.1
spacy==3.4.1
numpy==1.23.3
pandas==1.5.0
